Refine cleanup queue contract coverage
