+++ src/prefect/testing/standard_test_suites/__init__.py
@@ -1,1 +1,2 @@
 from .blocks import BlockStandardTestSuite
+from .worker_cleanup_queue import WorkerCleanupQueueStandardTestSuite
